Visualizing Spotify's Top Songs
Data visualization is a powerful tool to make numbers tell a story. The results of this work highlighted some interesting facts about what makes a song trendy, such as what type of music people listen to the most, or even how the high loudness and energy of a music often translate to a successful music.